feat: split serializers
