+++ iaso/api/microplanning/serializers/misc.py
@@ -2,7 +2,7 @@ from django.contrib.auth.models import User
 from drf_spectacular.utils import extend_schema_field, extend_schema_serializer
 from rest_framework import serializers
 
-from iaso.api.common import DateTimestampField, ModelSerializer, TimestampField
+from iaso.api.common import ModelSerializer, TimestampField
 from iaso.api.microplanning.filters import (
     validate_planning_has_org_unit_scope,
     validate_planning_org_unit_type_ids,
@@ -13,9 +13,7 @@ from iaso.models import (
     Form,
     Group,
     Mission,
-    MissionEntityType,
     MissionForm,
-    MissionOrgUnitType,
     OrgUnit,
     OrgUnitType,
     Planning,
@@ -23,7 +21,6 @@ from iaso.models import (
     Task,
 )
 from iaso.models.microplanning import Assignment, PlanningSamplingResult
-from iaso.models.missions import MissionType
 from iaso.models.org_unit import OrgUnitQuerySet
 from iaso.models.team import Team
 
@@ -75,181 +72,6 @@ class MissionFormSerializer(serializers.ModelSerializer):
         model = MissionForm
         fields = ["id", "form", "min_cardinality", "max_cardinality"]
         read_only_fields = ["id"]
-
-
-class MissionFormWriteSerializer(serializers.Serializer):
-    form_id = serializers.PrimaryKeyRelatedField(queryset=Form.objects.all(), source="form")
-    min_cardinality = serializers.IntegerField(default=1, min_value=0)
-    max_cardinality = serializers.IntegerField(required=False, allow_null=True, min_value=0)
-
-    def validate(self, attrs):
-        min_val = attrs.get("min_cardinality", 0)
-        max_val = attrs.get("max_cardinality")
-        if max_val is not None and min_val > max_val:
-            raise serializers.ValidationError({"min_cardinality": "min must be ≤ max"})
-        return attrs
-
-
-class MissionReadSerializer(serializers.ModelSerializer):
-    mission_forms = MissionFormSerializer(many=True, read_only=True)
-    org_unit_type = NestedOrgUnitTypeSerializer(read_only=True, source="org_unit_type.org_unit_type")
-    entity_type = NestedEntityTypeSerializer(read_only=True, source="entity_type.entity_type")
-    org_unit_min_cardinality = serializers.IntegerField(read_only=True, source="org_unit_type.min_cardinality")
-    org_unit_max_cardinality = serializers.IntegerField(read_only=True, source="org_unit_type.max_cardinality")
-    entity_min_cardinality = serializers.IntegerField(read_only=True, source="entity_type.min_cardinality")
-    entity_max_cardinality = serializers.IntegerField(read_only=True, source="entity_type.max_cardinality")
-
-    class Meta:
-        model = Mission
-        fields = [
-            "id",
-            "name",
-            "description",
-            "account",
-            "mission_type",
-            "mission_forms",
-            "org_unit_type",
-            "org_unit_min_cardinality",
-            "org_unit_max_cardinality",
-            "entity_type",
-            "entity_min_cardinality",
-            "entity_max_cardinality",
-            "created_by",
-            "created_at",
-            "updated_at",
-        ]
-        read_only_fields = fields
-
-
-class MissionWriteSerializer(serializers.ModelSerializer):
-    mission_forms = MissionFormWriteSerializer(many=True, required=False)
-    org_unit_type = serializers.PrimaryKeyRelatedField(
-        required=False, source="org_unit_type.org_unit_type", queryset=OrgUnitType.objects
-    )
-    org_unit_min_cardinality = serializers.IntegerField(required=False, source="org_unit_type.min_cardinality")
-    org_unit_max_cardinality = serializers.IntegerField(required=False, source="org_unit_type.max_cardinality")
-    entity_type = serializers.PrimaryKeyRelatedField(
-        required=False, source="entity_type.entity_type", queryset=EntityType.objects
-    )
-    entity_min_cardinality = serializers.IntegerField(required=False, source="entity_type.min_cardinality")
-    entity_max_cardinality = serializers.IntegerField(required=False, source="entity_type.max_cardinality")
-
-    class Meta:
-        model = Mission
-        fields = [
-            "id",
-            "name",
-            "description",
-            "mission_type",
-            "mission_forms",
-            "org_unit_type",
-            "org_unit_min_cardinality",
-            "org_unit_max_cardinality",
-            "entity_type",
-            "entity_min_cardinality",
-            "entity_max_cardinality",
-        ]
-        read_only_fields = ["id"]
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        user = self.context["request"].user
-        account = user.iaso_profile.account
-        self.fields["org_unit_type"].queryset = OrgUnitType.objects.filter(projects__account=account).distinct()
-        self.fields["entity_type"].queryset = EntityType.objects.filter(account=account)
-
-    def _validate_cardinality(self, object, errors, prefix):
-        min_val = object["min_cardinality"]
-        max_val = object["max_cardinality"]
-        if min_val is not None and max_val is not None and min_val > max_val:
-            errors[f"{prefix}_min_cardinality"] = "min must be ≤ max"
-
-    def validate(self, attrs):
-        validated_data = super().validate(attrs)
-        user = self.context["request"].user
-        validated_data["account"] = user.iaso_profile.account
-        validated_data["created_by"] = user
-        errors = {}
-        mission_type = validated_data.get("mission_type")
-
-        if mission_type == MissionType.FORM_FILLING:
-            if not validated_data.get("mission_forms"):
-                errors["mission_forms"] = "At least one form is required for FORM_FILLING missions."
-
-        elif mission_type == MissionType.ORG_UNIT_AND_FORM:
-            if not validated_data.get("org_unit_type"):
-                errors["org_unit_type"] = "Required for ORG_UNIT_AND_FORM missions."
-            else:
-                self._validate_cardinality(object=validated_data.get("org_unit_type"), errors=errors, prefix="org_unit")
-
-        elif mission_type == MissionType.ENTITY_AND_FORM:
-            if not validated_data.get("entity_type"):
-                errors["entity_type"] = "Required for ENTITY_AND_FORM missions."
-            else:
-                self._validate_cardinality(object=validated_data.get("entity_type"), errors=errors, prefix="entity")
-
-        if errors:
-            raise serializers.ValidationError(errors)
-
-        return validated_data
-
-    def _save_mission_forms(self, mission, mission_forms_data):
-        mission.mission_forms.all().delete()
-        for mf_data in mission_forms_data:
-            MissionForm.objects.create(
-                mission=mission,
-                form=mf_data["form"],
-                min_cardinality=mf_data.get("min_cardinality", 1),
-                max_cardinality=mf_data.get("max_cardinality"),
-            )
-
-    def create(self, validated_data):
-        mission_forms_data = validated_data.pop("mission_forms", [])
-        org_unit_type = validated_data.pop("org_unit_type", None)
-        entity_type = validated_data.pop("entity_type", None)
-        mission = super().create(validated_data)
-        if org_unit_type:
-            mission.org_unit_type = MissionOrgUnitType.objects.create(**org_unit_type)
-        if entity_type:
-            mission.entity_type = MissionEntityType.objects.create(**entity_type)
-        if org_unit_type or entity_type:
-            mission.save()
-        self._save_mission_forms(mission, mission_forms_data)
-        return mission
-
-    def update(self, instance, validated_data):
-        mission_forms_data = validated_data.pop("mission_forms", None)
-        org_unit_type = validated_data.pop("org_unit_type", None)
-        entity_type = validated_data.pop("entity_type", None)
-        mission = super().update(instance, validated_data)
-        updated = False
-        if org_unit_type:
-            if instance.org_unit_type:
-                instance.org_unit_type = MissionOrgUnitType.objects.update(**org_unit_type)
-            else:
-                mission.org_unit_type = MissionOrgUnitType.objects.create(**org_unit_type)
-                updated = True
-        elif instance.org_unit_type:
-            instance.org_unit_type.delete()
-        if entity_type:
-            if instance.entity_type:
-                instance.entity_type = MissionOrgUnitType.objects.update(**entity_type)
-            else:
-                mission.entity_type = MissionEntityType.objects.create(**entity_type)
-                updated = True
-        elif instance.entity_type:
-            instance.entity_type.delete()
-        if updated:
-            mission.save()
-        if mission_forms_data is not None:
-            self._save_mission_forms(mission, mission_forms_data)
-        return mission
-
-
-class AuditMissionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Mission
-        fields = "__all__"
 
 
 class NestedMissionSerializer(serializers.ModelSerializer):
@@ -722,106 +544,6 @@ class BulkDeleteAssignmentResponseSerializer(serializers.Serializer):
     deleted_count = serializers.IntegerField(read_only=True)
     planning_id = serializers.IntegerField(read_only=True)
     user = serializers.IntegerField(allow_null=True, read_only=True)
-
-
-# noinspection PyMethodMayBeStatic
-class MobilePlanningSerializer(serializers.ModelSerializer):
-    "Serialize plannings for mobile and transforms missions into forms to stay backward compatible with older versions"
-
-    def save(self):
-        # ensure that we can't save from here
-        raise NotImplementedError
-
-    created_at = TimestampField()
-    started_at = DateTimestampField()
-    ended_at = DateTimestampField()
-
-    assignments = serializers.SerializerMethodField()
-
-    class Meta:
-        model = Planning
-        fields = [
-            "id",
-            "name",
-            "description",
-            "created_at",
-            "started_at",
-            "ended_at",
-            "assignments",
-        ]
-
-    @staticmethod
-    def get_assignments(planning: Planning):
-        assignments = []
-        # Derive form_ids from missions for backward compatibility
-        planning_form_ids = set()
-        for mission in planning.missions.all():
-            planning_form_ids.update(mission.mission_forms.values_list("form_id", flat=True))
-
-        for a in planning.assignment_set.all():
-            out_set = set(a.org_unit.org_unit_type.form_set.values_list("id", flat=True))
-            intersection = out_set.intersection(planning_form_ids)
-            assignments.append({"org_unit_id": a.org_unit_id, "form_ids": intersection})
-        return assignments
-
-
-class MobilePlanningV2Serializer(serializers.ModelSerializer):
-    "Only used to serialize for mobile"
-
-    created_at = TimestampField()
-    started_at = DateTimestampField()
-    ended_at = DateTimestampField()
-    assignments = serializers.SerializerMethodField()
-
-    def save(self):
-        # ensure that we can't save from here
-        raise NotImplementedError
-
-    class Meta:
-        model = Planning
-        fields = [
-            "id",
-            "name",
-            "description",
-            "created_at",
-            "started_at",
-            "ended_at",
-            "assignments",
-        ]
-
-    @staticmethod
-    def get_assignments(planning: Planning):
-        assignments = []
-        for a in planning.assignment_set.all():
-            missions = []
-            for m in planning.missions.all():
-                if m.mission_type == MissionType.FORM_FILLING:
-                    # We assign the mission only if there is a match between the OUT's forms and the mission's forms
-                    out_set = set(a.org_unit.org_unit_type.form_set.values_list("id", flat=True))
-                    intersection = out_set.intersection(m.mission_forms.values_list("form_id", flat=True))
-                    if len(intersection) > 0:
-                        # Only keep the forms that are in the OUT
-                        mc = NestedMissionSerializer(m).data
-                        mc["mission_forms"] = list(
-                            filter(lambda x: x["form"]["id"] in intersection, mc["mission_forms"])
-                        )
-                        missions.append(mc)
-                elif m.mission_type == MissionType.ORG_UNIT_AND_FORM:
-                    # We need to filter on OrgUnit which are parent of the type
-                    m_out = m.org_unit_type.org_unit_type.id
-                    if (
-                        m_out == a.org_unit.org_unit_type.id
-                        or m_out in a.org_unit.org_unit_type.sub_unit_types.values_list("id", flat=True)
-                    ):
-                        missions.append(NestedMissionSerializer(m).data)
-                elif m.mission_type == MissionType.ENTITY_AND_FORM:
-                    # We always assign entities as there are no enforcement on entities and OrgUnit types.
-                    missions.append(NestedMissionSerializer(m).data)
-                else:
-                    raise NotImplementedError("Unknown mission type")
-            if len(missions) > 0:
-                assignments.append({"org_unit_id": a.org_unit_id, "missions": missions})
-        return assignments
 
 
 class PlanningOrgUnitSerializer(serializers.ModelSerializer):
