+++ iaso/api/microplanning/serializers/mobile/__init__.py
@@ -1,0 +1,5 @@
+from .v1 import MobilePlanningSerializer
+from .v2 import MobilePlanningV2Serializer
+
+
+__all__ = ["MobilePlanningSerializer", "MobilePlanningV2Serializer"]
